refactor SandboxDriver to use an AppState class per use case, plus other shared helpers
